laptop_scraping/requirements.txt
playwright
sqlalchemy
psycopg2-binary
python-dotenv
